- Created app directory and added first ui tests structure
