+++ requirements.txt
@@ -2,3 +2,5 @@ nicegui==2.24.0
 requests
 python-dotenv
 pytest
+pytest-asyncio
+pytest-selenium
