+++ tests/fixtures/base_customization.json
@@ -1,0 +1,26 @@
+{
+    "Color 1": "#f5f5f5",
+    "Color 2": "#e1e6ec",
+    "Color 3": "0055ff",
+    "Game Status Color": "#252525",
+    "Game Status Text Color": "#ffffff",
+    "Gradient": true,
+    "Height": 20,
+    "Left-Right": -19.5,
+    "Logos": true,
+    "Team 1 Color": "#919191",
+    "Team 1 Logo": "https://image.singular.live/7072b13f9e20b98034f48d6202400ff9/images/4trMUBAfgUDegi5GQFwydd_w400h267.png",
+    "Team 1 Logo Fit": "contain",
+    "Team 1 Text Color": "#ffffff",
+    "Team 1 Text Name": "Test Team A",
+    "Team 2 Color": "#015ecd",
+    "Team 2 Logo": "https://image.singular.live/7072b13f9e20b98034f48d6202400ff9/images/2TCM7xQ8puIGYr63pre3Y1_w200h105.png",
+    "Team 2 Logo Fit": "contain",
+    "Team 2 Text Color": "#ffffff",
+    "Team 2 Text Name": "TEST Team B",
+    "Text Color 1": "#0e0e0e",
+    "Text Color 2": "#0e0e0e",
+    "Text Color 3": "FFFFFF",
+    "Up-Down": "34",
+    "Width": 55
+  }
