+++ tests/fixtures/base_model.json
@@ -1,0 +1,17 @@
+{
+    "Serve": "None",
+    "Team 1 Game 1 Score": 0,
+    "Team 1 Game 2 Score": "0",
+    "Team 1 Game 3 Score": "0",
+    "Team 1 Game 4 Score": "0",
+    "Team 1 Game 5 Score": "0",
+    "Team 1 Sets": "0",
+    "Team 1 Timeouts": "0",
+    "Team 2 Game 1 Score": 0,
+    "Team 2 Game 2 Score": "0",
+    "Team 2 Game 3 Score": "0",
+    "Team 2 Game 4 Score": "0",
+    "Team 2 Game 5 Score": "0",
+    "Team 2 Sets": "0",
+    "Team 2 Timeouts": "0"
+}
