+++ tests/fixtures/midgame_model.json
@@ -1,0 +1,17 @@
+{
+    "Serve": "A",
+    "Team 1 Game 1 Score": 16,
+    "Team 1 Game 2 Score": 25,
+    "Team 1 Game 3 Score": 11,
+    "Team 1 Game 4 Score": "0",
+    "Team 1 Game 5 Score": "0",
+    "Team 1 Sets": 1,
+    "Team 1 Timeouts": "0",
+    "Team 2 Game 1 Score": 25,
+    "Team 2 Game 2 Score": 14,
+    "Team 2 Game 3 Score": 9,
+    "Team 2 Game 4 Score": "0",
+    "Team 2 Game 5 Score": "0",
+    "Team 2 Sets": 1,
+    "Team 2 Timeouts": "1"
+  }
