+++ tests/test_state.py
@@ -6,7 +6,7 @@ import copy
 # Add the project's root directory to the Python path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from state import State
+from app.state import State
 
 @pytest.fixture
 def state():
